% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-methods.R
\name{compare_test}
\alias{compare_test}
\title{Compare MEQLEA to existing methods}
\usage{
compare_test(dat)
}
\arguments{
\item{num_gene}{number of genes to be simulated.}

\item{prop}{a vector of length 2, corresponding to $p_t$ and $p_b$.}

\item{delta}{a vector of length num_gene, representing DE effect for each gene}

\item{case}{which case to simulate (see simulation setup in the paper)}

\item{rho}{a vector of length 3. Corresponds \eqn{\rho_1}, \eqn{\rho_2} and \eqn{\rho_3}.}
}
\value{
a list 
\item{go_term}{ DE indicator for each simulated gene. 1 is for genes in the test set, and 0 othwerwise.}
\item{delta}{DE effect for each simulated gene.}
\item{sigma}{the true correlation, used to simulate  the expression data}
}
\description{
Compare MEQLEA to existing methods.
}

