% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meqlea.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{Estimate sample covariance.}
\usage{
estimate_sigma(expression_data, trt)
}
\arguments{
\item{obj}{object returned by \code{prepare_simulation}.}

\item{size}{number of samples to be simulated}
}
\value{
a list 
\item{data}{a expression matrix of \eqn{m\times n} where m is the number of genes and n is the number of samples. }
\item{trt}{sample labels of length \code{n}, 1 for treatment and 0 for control.}
\item{go_term}{gene labels of length \code{m}, 1 for go_term genes and 0 otherwise.}
\item{sigma}{true covariance matrix upon which data is simulated.}
}
\description{
Estimate sample covariance and calculate the gene-level statistics
}

