% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meqlea.R
\name{meqlea_multiple}
\alias{meqlea_multiple}
\title{MEQLEA-multiple.}
\usage{
meqlea_multiple(expression_data, trt, geneset, standardize = T,
  minSetSize = 5, fdr_method = "BH")
}
\arguments{
\item{expression_data}{the expressoin matrix.}

\item{trt}{treatment labels.}

\item{geneset}{gene sets to be tested, an object from \code{read_gene_set}.}

\item{standardize}{whether the data should be standaridzed.}

\item{minSetSize}{the minimum number of genes contained for a gene set to be considered.}

\item{fdr_method}{which method is ued to adjust the p values. see arguments in function \code{p.adjust}.}
}
\value{
a data frame
}
\description{
meqlea for testing multiple gene sets.
}

