% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meqlea.R
\name{meqlea_single}
\alias{meqlea_single}
\title{MEQLEA-single.}
\usage{
meqlea_single(expression_data, trt, go_term, standardize = F)
}
\arguments{
\item{expression_data}{the expressoin matrix.}

\item{trt}{treatment labels.}

\item{go_term}{an indicator vector. 1 for genes in the test, 0 otherwise.}

\item{standardize}{whether the data should be standaridzed.}
}
\value{
a list 
\item{set_name}{The name of the gene set}
\item{testSetCor}{Average correlation for genes in the test set}
\item{interCor}{Average correlation between genes in the test set and those not in the test set}
\item{backSetCor}{Average correlations for genes not in the test set.}
}
\description{
meqlea for single gene set test.
}

