% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gene-set.R
\name{GSEA.Gct2Frame2}
\alias{GSEA.Gct2Frame2}
\title{Convert gene sets to lists}
\usage{
GSEA.Gct2Frame2(filename = "NULL")
}
\arguments{
\item{file}{the expression data files.}
}
\description{
read the expression data of Gct format.
}

